Move tool implementation to subclasses
